Feat: Improved dbServer exception handling and reporting.
Feat: Added improved validation for MasterUser, now checks for already created users and validates username and password before trying to query the server.

Feat: When exceptions occur on dBServer, the exception is now integrated in the ServerResponse with the included Exception message - which can be shown to the user at higher levels if appropriate.